Update solution for sigmoid-numpy

+++ sigmoid-numpy/sigmoid-numpy.py
@@ -5,5 +5,5 @@ def sigmoid(x):
     Vectorized sigmoid function.
     """
     # Write code here
-    x=np.asarray(x)
-    return 1.0/(1.0+np.exp(-x))
+    x=np.asarray(x,dtype=float)
+    return 1/(1+np.exp(-x))
